Add files via upload
This program models the ignition system of a standard gasoline based car. It is initially set to run. The user can open the source code and alter the variable. For example, if you lower the battery voltage below the threshold, the program will notify you that you need to charge your battery. Don't forget the fuel!